#14 extend selection support
Selection now supports <, <=, >, >= in addition to ==
