+++ internal/compiler/optimization/half_join_bench_test.go
@@ -12,10 +12,10 @@ func Benchmark_OptHalfJoin(b *testing.B) {
 	cmd := command.Project{
 		Cols: []command.Column{
 			{
-				Column: command.LiteralExpr{Value: "col1"},
-				Alias:  "myCol",
+				Name:  command.LiteralExpr{Value: "col1"},
+				Alias: "myCol",
 			},
-			{Column: command.LiteralExpr{Value: "col2"}},
+			{Name: command.LiteralExpr{Value: "col2"}},
 		},
 		Input: command.Select{
 			Filter: command.EqualityExpr{
