+++ internal/compiler/testdata/TestCompileGolden/select/#03.golden
@@ -1,4 +1,4 @@
-command.Limit{Limit:command.LiteralExpr{Value:"5"}, Input:command.Offset{Offset:command.LiteralExpr{Value:"10"}, Input:command.Project{Cols:[]command.Column{command.Column{Table:"", Column:command.LiteralExpr{Value:"*"}, Alias:""}}, Input:command.Scan{Table:command.SimpleTable{Schema:"", Table:"myTable", Alias:"", Indexed:false, Index:""}}}}}
+command.Limit{Limit:command.LiteralExpr{Value:"5"}, Input:command.Offset{Offset:command.LiteralExpr{Value:"10"}, Input:command.Project{Cols:[]command.Column{command.Column{Table:"", Name:command.LiteralExpr{Value:"*"}, Alias:""}}, Input:command.Scan{Table:command.SimpleTable{Schema:"", Table:"myTable", Alias:"", Indexed:false, Index:""}}}}}
 
 String:
 Limit[limit=5](Offset[offset=10](Project[cols=*](Scan[table=myTable]())))
