+++ internal/compiler/testdata/TestCompileGolden/select/#05.golden
@@ -1,4 +1,4 @@
-command.Distinct{Input:command.Project{Cols:[]command.Column{command.Column{Table:"", Column:command.LiteralExpr{Value:"*"}, Alias:""}}, Input:command.Select{Filter:command.ConstantBooleanExpr{Value:true}, Input:command.Scan{Table:command.SimpleTable{Schema:"", Table:"myTable", Alias:"", Indexed:false, Index:""}}}}}
+command.Distinct{Input:command.Project{Cols:[]command.Column{command.Column{Table:"", Name:command.LiteralExpr{Value:"*"}, Alias:""}}, Input:command.Select{Filter:command.ConstantBooleanExpr{Value:true}, Input:command.Scan{Table:command.SimpleTable{Schema:"", Table:"myTable", Alias:"", Indexed:false, Index:""}}}}}
 
 String:
 Distinct[](Project[cols=*](Select[filter=true](Scan[table=myTable]())))
