+++ internal/engine/evaluate_test.go
@@ -6,7 +6,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/xqueries/xdb/internal/compiler/command"
 	"github.com/xqueries/xdb/internal/engine/table"
-	"github.com/xqueries/xdb/internal/engine/types"
 )
 
 func TestFullTableScan(t *testing.T) {
@@ -22,326 +21,4 @@ func TestFullTableScan(t *testing.T) {
 	})
 	assert.NoError(err)
 	assert.Equal(table.Table{}, result)
-}
-
-func TestEngine_evaluateProjection(t *testing.T) {
-	tests := []struct {
-		name    string
-		ctx     ExecutionContext
-		proj    command.Project
-		want    table.Table
-		wantErr string
-	}{
-		{
-			"empty",
-			newEmptyExecutionContext(),
-			command.Project{
-				Cols: []command.Column{},
-				Input: command.Values{
-					Values: [][]command.Expr{
-						{command.LiteralExpr{Value: "hello"}, command.LiteralExpr{Value: "world"}, command.ConstantBooleanExpr{Value: true}},
-						{command.LiteralExpr{Value: "foo"}, command.LiteralExpr{Value: "bar"}, command.ConstantBooleanExpr{Value: false}},
-					},
-				},
-			},
-			table.Table{
-				Cols: []table.Col{},
-				Rows: []table.Row{},
-			},
-			"",
-		},
-		{
-			"simple",
-			newEmptyExecutionContext(),
-			command.Project{
-				Cols: []command.Column{
-					{
-						Column: command.LiteralExpr{Value: "column2"},
-					},
-				},
-				Input: command.Values{
-					Values: [][]command.Expr{
-						{command.LiteralExpr{Value: "hello"}, command.LiteralExpr{Value: "world"}, command.ConstantBooleanExpr{Value: true}},
-						{command.LiteralExpr{Value: "foo"}, command.LiteralExpr{Value: "bar"}, command.ConstantBooleanExpr{Value: false}},
-					},
-				},
-			},
-			table.Table{
-				Cols: []table.Col{
-					{
-						QualifiedName: "column2",
-						Type:          types.String,
-					},
-				},
-				Rows: []table.Row{
-					{
-						Values: []types.Value{types.NewString("world")},
-					},
-					{
-						Values: []types.Value{types.NewString("bar")},
-					},
-				},
-			},
-			"",
-		},
-		{
-			"simple with alias",
-			newEmptyExecutionContext(),
-			command.Project{
-				Cols: []command.Column{
-					{
-						Column: command.LiteralExpr{Value: "column2"},
-						Alias:  "foo",
-					},
-				},
-				Input: command.Values{
-					Values: [][]command.Expr{
-						{command.LiteralExpr{Value: "hello"}, command.LiteralExpr{Value: "world"}, command.ConstantBooleanExpr{Value: true}},
-						{command.LiteralExpr{Value: "foo"}, command.LiteralExpr{Value: "bar"}, command.ConstantBooleanExpr{Value: false}},
-					},
-				},
-			},
-			table.Table{
-				Cols: []table.Col{
-					{
-						QualifiedName: "column2",
-						Alias:         "foo",
-						Type:          types.String,
-					},
-				},
-				Rows: []table.Row{
-					{
-						Values: []types.Value{types.NewString("world")},
-					},
-					{
-						Values: []types.Value{types.NewString("bar")},
-					},
-				},
-			},
-			"",
-		},
-		{
-			"missing column",
-			newEmptyExecutionContext(),
-			command.Project{
-				Cols: []command.Column{
-					{
-						Column: command.LiteralExpr{Value: "foo"},
-					},
-				},
-				Input: command.Values{
-					Values: [][]command.Expr{
-						{command.LiteralExpr{Value: "hello"}},
-						{command.LiteralExpr{Value: "foo"}},
-					},
-				},
-			},
-			table.Table{},
-			"no column with name or alias 'foo'",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert := assert.New(t)
-
-			e := createEngineOnEmptyDatabase(t)
-			got, err := e.evaluateProjection(tt.ctx, tt.proj)
-			if tt.wantErr != "" {
-				assert.EqualError(err, tt.wantErr)
-			} else {
-				assert.NoError(err)
-			}
-			assert.Equal(tt.want, got)
-		})
-	}
-}
-
-func TestEngine_evaluateSelection(t *testing.T) {
-	tests := []struct {
-		name    string
-		ctx     ExecutionContext
-		sel     command.Select
-		want    table.Table
-		wantErr string
-	}{
-		{
-			"trivial",
-			newEmptyExecutionContext(),
-			command.Select{
-				Filter: command.ConstantBooleanExpr{Value: true},
-				Input: command.Values{
-					Values: [][]command.Expr{
-						{command.LiteralExpr{Value: "hello"}, command.LiteralExpr{Value: "5"}, command.ConstantBooleanExpr{Value: true}},
-						{command.LiteralExpr{Value: "foo"}, command.LiteralExpr{Value: "7"}, command.ConstantBooleanExpr{Value: false}},
-					},
-				},
-			},
-			table.Table{
-				Cols: []table.Col{
-					{
-						QualifiedName: "column1",
-						Type:          types.String,
-					},
-					{
-						QualifiedName: "column2",
-						Type:          types.Integer,
-					},
-					{
-						QualifiedName: "column3",
-						Type:          types.Bool,
-					},
-				},
-				Rows: []table.Row{
-					{
-						Values: []types.Value{types.NewString("hello"), types.NewInteger(5), types.NewBool(true)},
-					},
-					{
-						Values: []types.Value{types.NewString("foo"), types.NewInteger(7), types.NewBool(false)},
-					},
-				},
-			},
-			"",
-		},
-		{
-			"simple",
-			newEmptyExecutionContext(),
-			command.Select{
-				Filter: command.EqualityExpr{
-					BinaryBase: command.BinaryBase{
-						Left:  command.LiteralExpr{Value: "column2"},
-						Right: command.LiteralExpr{Value: "7"},
-					},
-				},
-				Input: command.Values{
-					Values: [][]command.Expr{
-						{command.LiteralExpr{Value: "hello"}, command.LiteralExpr{Value: "5"}, command.ConstantBooleanExpr{Value: true}},
-						{command.LiteralExpr{Value: "foo"}, command.LiteralExpr{Value: "7"}, command.ConstantBooleanExpr{Value: false}},
-					},
-				},
-			},
-			table.Table{
-				Cols: []table.Col{
-					{
-						QualifiedName: "column1",
-						Type:          types.String,
-					},
-					{
-						QualifiedName: "column2",
-						Type:          types.Integer,
-					},
-					{
-						QualifiedName: "column3",
-						Type:          types.Bool,
-					},
-				},
-				Rows: []table.Row{
-					{
-						Values: []types.Value{types.NewString("foo"), types.NewInteger(7), types.NewBool(false)},
-					},
-				},
-			},
-			"",
-		},
-		{
-			"erronous filter",
-			newEmptyExecutionContext(),
-			command.Select{
-				Filter: command.LiteralExpr{Value: "erronous"},
-				Input: command.Values{
-					Values: [][]command.Expr{
-						{command.LiteralExpr{Value: "hello"}, command.LiteralExpr{Value: "5"}, command.ConstantBooleanExpr{Value: true}},
-						{command.LiteralExpr{Value: "foo"}, command.LiteralExpr{Value: "7"}, command.ConstantBooleanExpr{Value: false}},
-					},
-				},
-			},
-			table.Table{},
-			"cannot use command.LiteralExpr as filter",
-		},
-		{
-			"column against column",
-			newEmptyExecutionContext(),
-			command.Select{
-				Filter: command.EqualityExpr{
-					BinaryBase: command.BinaryBase{
-						Left:  command.LiteralExpr{Value: "column2"},
-						Right: command.LiteralExpr{Value: "column1"},
-					},
-				},
-				Input: command.Values{
-					Values: [][]command.Expr{
-						{command.LiteralExpr{Value: "hello"}, command.LiteralExpr{Value: "world"}},
-						{command.LiteralExpr{Value: "foo"}, command.LiteralExpr{Value: "foo"}},
-					},
-				},
-			},
-			table.Table{
-				Cols: []table.Col{
-					{
-						QualifiedName: "column1",
-						Type:          types.String,
-					},
-					{
-						QualifiedName: "column2",
-						Type:          types.String,
-					},
-				},
-				Rows: []table.Row{
-					{
-						Values: []types.Value{types.NewString("foo"), types.NewString("foo")},
-					},
-				},
-			},
-			"",
-		},
-		{
-			"column against string",
-			newEmptyExecutionContext(),
-			command.Select{
-				Filter: command.EqualityExpr{
-					BinaryBase: command.BinaryBase{
-						Left:  command.LiteralExpr{Value: "column2"},
-						Right: command.LiteralExpr{Value: "world"},
-					},
-				},
-				Input: command.Values{
-					Values: [][]command.Expr{
-						{command.LiteralExpr{Value: "hello"}, command.LiteralExpr{Value: "world"}},
-						{command.LiteralExpr{Value: "foo"}, command.LiteralExpr{Value: "foo"}},
-					},
-				},
-			},
-			table.Table{
-				Cols: []table.Col{
-					{
-						QualifiedName: "column1",
-						Type:          types.String,
-					},
-					{
-						QualifiedName: "column2",
-						Type:          types.String,
-					},
-				},
-				Rows: []table.Row{
-					{
-						Values: []types.Value{types.NewString("hello"), types.NewString("world")},
-					},
-				},
-			},
-			"",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			assert := assert.New(t)
-
-			e := createEngineOnEmptyDatabase(t)
-			got, err := e.evaluateSelection(tt.ctx, tt.sel)
-			if tt.wantErr != "" {
-				assert.EqualError(err, tt.wantErr)
-			} else {
-				assert.NoError(err)
-			}
-			assert.Equal(tt.want, got)
-		})
-	}
 }
