+++ internal/test/testdata/example05/output
@@ -1,0 +1,3 @@
+leftCol (Integer)   rightCol (Integer)
+3                   6
+5                   9
